@@ -1,47 +1,60 @@
 {
-    "name": "togather-front",
-    "version": "0.1.0",
-    "private": true,
-    "dependencies": {
-        "@emotion/react": "^11.11.3",
-        "@emotion/styled": "^11.11.0",
-        "@testing-library/jest-dom": "^5.17.0",
-        "@testing-library/react": "^13.4.0",
-        "@testing-library/user-event": "^13.5.0",
-        "react": "17.0.2",
-        "react-dom": "17.0.2",
-        "react-router-dom": "^6.22.1",
-        "react-scripts": "5.0.1",
-        "react-virtualize": "^1.2.9",
-        "web-vitals": "^2.1.4"
-    },
-    "scripts": {
-        "start": "react-scripts start",
-        "build": "react-scripts build",
-        "test": "react-scripts test",
-        "eject": "react-scripts eject"
-    },
-    "eslintConfig": {
-        "extends": [
-            "react-app",
-            "react-app/jest"
-        ]
-    },
-    "browserslist": {
-        "production": [
-            ">0.2%",
-            "not dead",
-            "not op_mini all"
-        ],
-        "development": [
-            "last 1 chrome version",
-            "last 1 firefox version",
-            "last 1 safari version"
-        ]
-    },
-    "devDependencies": {
-        "eslint-config-prettier": "^9.1.0",
-        "eslint-plugin-prettier": "^5.1.3",
-        "prettier": "^3.2.5"
-    }
+  "name": "togather-front",
+  "version": "0.1.0",
+  "private": true,
+  "dependencies": {
+    "@emotion/react": "^11.11.3",
+    "@emotion/styled": "^11.11.0",
+    "@testing-library/jest-dom": "^5.17.0",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^13.5.0",
+    "@tosspayments/payment-widget-sdk": "^0.10.2",
+    "axios": "^1.6.8",
+    "moment": "^2.30.1",
+    "react": "^18.2.0",
+    "react-calendar": "^4.8.0",
+    "react-daum-postcode": "^3.1.3",
+    "react-dom": "^18.2.0",
+    "react-icons": "^5.0.1",
+    "react-kakao-maps-sdk": "^1.1.26",
+    "react-router-dom": "^6.22.3",
+    "react-scripts": "5.0.1",
+    "react-spinners": "^0.13.8",
+    "react-virtualized": "^9.22.5",
+    "recoil": "^0.7.7",
+    "recoil-persist": "^5.1.0",
+    "styled-normalize": "^8.1.1",
+    "styled-reset": "^4.5.2",
+    "web-vitals": "^2.1.4"
+  },
+  "scripts": {
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject"
+  },
+  "eslintConfig": {
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
+  },
+  "devDependencies": {
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-prettier": "^5.1.3",
+    "prettier": "^3.2.5"
+  },
+  "proxy": "http://localhost:8080/"
 }
